CORE-6119 Rename libs/crypto/merkle to libs/crypto/merkle-impl to avoid clashing with its API counterpart
