+++ libs/crypto/merkle-impl/test.bndrun
@@ -5,7 +5,7 @@
 -runtrace: true
 
 -runrequires: \
-    bnd.identity;id='net.corda.merkle',\
+    bnd.identity;id='net.corda.merkle-impl',\
     bnd.identity;id='${project.archivesBaseName}-tests',\
     bnd.identity;id='junit-jupiter-engine',\
     bnd.identity;id='junit-platform-launcher',\
