+++ settings.gradle
@@ -196,7 +196,7 @@ include 'libs:crypto:crypto-config-impl'
 include 'libs:crypto:crypto-core'
 include 'libs:crypto:crypto-flow'
 include 'libs:crypto:crypto-impl'
-include 'libs:crypto:merkle'
+include 'libs:crypto:merkle-impl'
 include 'libs:crypto:crypto-serialization-impl'
 include 'libs:crypto:delegated-signing'
 include 'libs:crypto:certificate-generation'
